requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
pystray>=0.19.0
Pillow>=10.0.0
flask-socketio>=5.3.0
python-socketio>=5.11.0
PyJWT>=2.8.0
